Modified answer range from int to long long

+++ 04-Binary-Search/02-BS-On-Answers/01-Find-Square-Root-Of-A-Number.cpp
@@ -14,7 +14,7 @@ public:
 
         while (low <= high)
         {
-            int mid = low + (high - low) / 2;
+            long long mid = low + (high - low) / 2;
 
             if (mid * mid <= number)
             {
